mtu-migration-dispatcher script
A new Network Manager dispatcher script, placed in mtu-migration-dispatcher.yaml, responds to UP events on selected host interfaces when an MTU migration is ongoing. It sets appropriate (transient) MTU values for host interfaces and their routes in the first phase of the MTU migration in order to avoid disruption while the cluster MTU is also being modified.

Permanently applying an MTU value to host interfaces is instead under the cluster administrators's responsibility; an option for the cluster admin, for instance, would be to apply a MachineConfig object that adds a custom Network Manager profile to the host.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>